feat: add router, error page
