+++ package.json
@@ -15,7 +15,8 @@
 	"dependencies": {
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
-		"react-error-boundary": "^4.0.13"
+		"react-error-boundary": "^4.0.13",
+		"react-router-dom": "^6.24.1"
 	},
 	"devDependencies": {
 		"@types/node": "^20.14.10",
@@ -28,6 +29,8 @@
 		"eslint-config-airbnb": "^19.0.4",
 		"eslint-config-airbnb-typescript": "^18.0.0",
 		"eslint-config-prettier": "^9.1.0",
+		"eslint-import-resolver-alias": "^1.1.2",
+		"eslint-import-resolver-typescript": "^3.6.1",
 		"eslint-plugin-prettier": "^5.1.3",
 		"eslint-plugin-react": "^7.34.3",
 		"eslint-plugin-react-compiler": "^0.0.0-experimental-0998c1e-20240625",
