+++ src/pages/error/ErrorPage.module.scss
@@ -1,0 +1,31 @@
+@import "@styles/variables", "@styles/mixins";
+
+.wrapper {
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	height: 80vh;
+}
+
+.pikachu {
+	width: 400px;
+	height: 380px;
+}
+
+.content {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	gap: 5px;
+
+	h2 {
+		font-size: 3rem;
+		margin: 0%;
+	}
+
+	button {
+		@include button-style;
+		background-color: $button-bg-color;
+		color: $button-color;
+	}
+}
