+++ src/pages/home/HomePage.module.scss
@@ -8,8 +8,8 @@
 	justify-content: center;
 	align-items: center;
 
-	.logo-container,
-	.search-container {
+	.logoContainer,
+	.searchContainer {
 		display: flex;
 		justify-content: center;
 		align-items: center;
@@ -22,7 +22,7 @@
 		animation: spin 8s linear infinite;
 	}
 
-	.logo-text {
+	.logoText {
 		display: flex;
 		flex-direction: column;
 		justify-content: center;
@@ -37,7 +37,7 @@
 		}
 	}
 
-	.preview-cards-container {
+	.previewCardsContainer {
 		display: flex;
 		justify-content: center;
 		flex-wrap: wrap;
