+++ src/router/index.tsx
@@ -1,0 +1,21 @@
+import { createBrowserRouter } from "react-router-dom";
+
+import App from "@/App";
+import HomePage from "@/pages/home/HomePage";
+import ErrorPage from "@/pages/error/ErrorPage";
+
+const router = createBrowserRouter([
+	{
+		path: "/",
+		element: <App />,
+		errorElement: <ErrorPage />,
+		children: [
+			{
+				path: "/",
+				element: <HomePage />,
+			},
+		],
+	},
+]);
+
+export default router;
