+++ tsconfig.json
@@ -24,6 +24,8 @@
 		"noFallthroughCasesInSwitch": true,
 		"baseUrl": ".",
 		"paths": {
+			"@/*": ["src/*"],
+			"@components/*": ["src/components/*"],
 			"@styles/*": ["src/styles/*"]
 		}
 	},
