+++ vite.config.ts
@@ -7,6 +7,8 @@ export default defineConfig({
 	plugins: [react()],
 	resolve: {
 		alias: {
+			"@": path.resolve(__dirname, "src"),
+			"@components": path.resolve(__dirname, "src/components"),
 			"@styles": path.resolve(__dirname, "src/styles"),
 		},
 	},
